support our 2 far more complex call structures
these are arrays at run time; the one is then converted into a
drush_invoke (not a big deal) the other kicks off a recursive thread to
issue another recipe be executed since it’s a conditional as to which
they can select. once that thread finishes the original one will resume
so you can nest conditionals / branching paths which means we’ll really
need to document this fact.